Accepts a list of predefined repos as parameter
First it populates the list based on the repos found in the
repo_root directory specified as parameter.

Later, when calling the endpoint, the name of the repo is checked
against that repo list and it will only accept requests with a
valid repository as parameter.

This commit should fix the vulnerability mentioned in #12.

Still need to add some tests on the code and some other stuff
mentioned as TODOs.

Signed-off-by: Albert Pastrana <albert.pastrana@intenthq.com>