Don't rely on local user and email for Git tests.
On Travis, depending on what OS you are building on, this may or may not
be set — and I suppose it's infinitely possible this could be the case
on a user's machine too! So, instead, we now set the signature (that is,
email, user's name and time) of the commit to a statically defined thing
we control in the tests.

This fixes the test failures we see on at least MacOS on Travis CI.