Allow to `ls` a directory.
Add a new endpoint to `ls` a directory to see all the files underneath.
The files are returned as a JSON array of strings representing the
relative paths of the files underneath the given directory.

The result is non-recursive because otherwise there is an opportunity to
DOS the server with a large response on a deep directory tree — it seems
more unlikely that there is a directory with children directly under it
large enough to break this but of course the risk there is also
non-zero.

The endpoint is `/repos/<name>/ls/<path>`.